Allow regenerating embeddings (#15)
This adds a command to allow regenerating embeddings. At the same time,
to improve search results, we now use OpenAI to visually analyse the
image for faces and get a description of the face. We then generate the
description for each.

To further refine search results, we increase the number of candidates
to twenty times the number of results to be returned. This is suggested
in the [$vectorSearch
docs](https://www.mongodb.com/docs/atlas/atlas-vector-search/vector-search-stage/#std-label-avs-num-candidates):
> For ANN search, we recommend that you specify a `numCandidates` number
at least 20 times higher than the number of documents to return
(`limit`) to increase accuracy.