Add event listener to fix vector search indexes (#16)
Since the `SearchIndex` mapping attribute doesn't support vector search
indexes yet, this PR adds an event listener to work around the issue.
This listener handles the `loadClassMetadata` event and reshapes the
index definition of the search index to match the vector search format.
Specifically, we set the `type` attribute and extract the `fields` from
the `mappings` definition, as this differs from text search indexes.

Note that this PR completely ignores the `readonly` designation (through
a docblock) of the `searchIndexes` field in ClassMetadata, but I hope
that we can add support for vector search indexes to ODM soon.