Add match page (#19)
The new match page shows the original picture side-by-side with the most
likely match. To produce better results, we now take the three most
likely matches from our vector search and pass those on to OpenAI to let
facial analysis do the rest. To reduce the number of AI requests, the
most similar result is now stored with the face. This also helps with
avoiding inconsistent results.

Note: I have not yet changed the webcam page to redirect to the new
match page, but IMO we should do that as it produces a better result.